postgres_to_es/requirements.txt
pydantic==1.10.7
psycopg2-binary==2.9.6
python-dotenv==1.0.0
elasticsearch==8.8.0
backoff==2.2.1